requirements.txt
meteomatics>=2.9.0
pandas>=1.4.3
SQLAlchemy>=2.0.7
python-dateutil>=2.8.1
fastapi>=0.94.1
psycopg2>=2.9.5
psycopg2-binary